requirements.txt
python-dotenv
telethon
psycopg2
pandas>=1.1.4
seaborn>=0.11.0
sqlalchemy
dbt-core
dbt-postgres
emoji
opencv-python
Pillow
gitpython>=3.1.30
matplotlib>=3.3
numpy>=1.23.5
opencv-python>=4.1.1
pillow>=10.3.0
psutil  # system resources
PyYAML>=5.3.1
requests>=2.32.2
scipy>=1.4.1
thop>=0.1.1  # FLOPs computation
torch>=1.8.0  # see https://pytorch.org/get-started/locally (recommended)
torchvision>=0.9.0
tqdm>=4.66.3
ultralytics>=8.2.34 
setuptools>=70.0.0
